list command lists all accounts in all namespaces by default
If no namespace is provided explicitly, for both backends, all service
accounts in all namespaces that have the custom labels provided by this
tool should be listed. Earlier, only the accounts in the `default`
namespace were listed when no `--namespace` option was provided. This is
done for the `lightkube` backend to be consistent with the `kubectl`
backend, because `kubectl` lists all accounts in all namespaces by
default.