Check if the resources already exists before creating service account.
This needs to be done to ensure the service account, role and role
binding to be created as part of creation of service account do not
exist already in the Kubernetes cluster. If such scenario happens, then
the program would crash with exception, while it may have created some
resources already but still not able to create all of the resources.